# SafeNotes.Angular/openapi.yaml
openapi: 3.0.1
info:
  title: Demo API
  version: v1
paths:
  /api/Auth:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtTokenResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/JwtTokenResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/JwtTokenResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/Notes:
    post:
      tags:
        - Notes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/IdModel'
            application/json:
              schema:
                $ref: '#/components/schemas/IdModel'
            text/json:
              schema:
                $ref: '#/components/schemas/IdModel'
  /api/Notes/{id}:
    put:
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '204':
          description: Success
  /api/Notes/list:
    get:
      tags:
        - Notes
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetNoteListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetNoteListResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetNoteListResponse'
  /api/Notes/{id}/access:
    get:
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetAllowedUsersResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllowedUsersResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetAllowedUsersResponse'
    put:
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAllowedUsersRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateAllowedUsersRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateAllowedUsersRequest'
      responses:
        '201':
          description: Success
  /api/Notes/get/{id}:
    put:
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetNoteRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/GetNoteRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/GetNoteRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetNoteResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetNoteResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetNoteResponse'
  /api/Users:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '200':
          description: Success
  /api/Users/confirm-email/{emailConfirmationCode}:
    get:
      tags:
        - Users
      parameters:
        - name: emailConfirmationCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
components:
  schemas:
    AllowedUser:
      type: object
      properties:
        email:
          type: string
          nullable: true
        canRead:
          type: boolean
          nullable: true
        canWrite:
          type: boolean
          nullable: true
      additionalProperties: false
    CreateNoteRequest:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        isPublic:
          type: boolean
          nullable: true
        isEncryptedWithUserSpecifiedKey:
          type: boolean
          nullable: true
        key:
          type: string
          nullable: true
      additionalProperties: false
    GetAllowedUsersResponse:
      type: object
      properties:
        allowedUsers:
          type: array
          items:
            $ref: '#/components/schemas/AllowedUser'
          nullable: true
      additionalProperties: false
    GetNoteListResponse:
      type: object
      properties:
        userNotes:
          type: array
          items:
            $ref: '#/components/schemas/NotePreview'
          nullable: true
        sharedToUserNotes:
          type: array
          items:
            $ref: '#/components/schemas/NotePreview'
          nullable: true
      additionalProperties: false
    GetNoteRequest:
      type: object
      properties:
        key:
          type: string
          nullable: true
      additionalProperties: false
    GetNoteResponse:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        isPublic:
          type: boolean
        isEncryptedWithUserSpecifiedKey:
          type: boolean
        lastModifiedOn:
          type: string
          format: date-time
      additionalProperties: false
    IdModel:
      type: object
      properties:
        id:
          type: integer
          format: int32
      additionalProperties: false
    JwtTokenResponse:
      type: object
      properties:
        token:
          type: string
          nullable: true
      additionalProperties: false
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    NotePreview:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
          nullable: true
        isPublic:
          type: boolean
        isEncryptedWithUserSpecifiedKey:
          type: boolean
          nullable: true
        lastModified:
          type: string
          format: date-time
        canWrite:
          type: boolean
        isOwner:
          type: boolean
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
    RegisterUserRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    UpdateAllowedUsersRequest:
      type: object
      properties:
        noteId:
          type: integer
          format: int32
        allowedUsers:
          type: array
          items:
            $ref: '#/components/schemas/AllowedUser'
          nullable: true
      additionalProperties: false
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        key:
          type: string
          nullable: true
      additionalProperties: false
